Complete flex exercise 4 of TOP
I completed the 4th flex exercise of TOP which focused on designing
an information page. I learned that text-align can apply to images as
well if they are placed inside a block element. Before I knew this, I
used flex on the info cards just to center the image really which was a
waste seeing as I could make use text-align to center text and images.