Complete flex exercise 5 TOP
I completed the 5th flex exercise of TOP. It involved practicing
creating and alert box. My solution worked but it was a little different
to TOP's solution.

The solution I implemented involved creating a container that wrapped
and allowed the heading and close button to be on the top and the text
to wrap below this. I then used auto-margin to move the close button to
the right.

TOP's solution put the heading and close button into one flex container,
which allowed them to work on moving the heading and close button and
allow the text to move below it by default.